TG/coruja/test_requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
httpx==0.25.2 